requirements/base.txt
celery==4.0.2
dj-static==0.0.6
django-debug-toolbar==1.7.0
django-filter==1.0.2
django-health-check==2.3.0
django-model-utils==3.0.0
django-redis==4.8
django-rest-swagger==2.1.2
django-timezone-field==2.0rc1
Django==1.10.5
djangorestframework==3.6.2
gunicorn==19.7.1
moment==0.6.1
psycopg2==2.7.1
python-dateutil==2.6.0
social-auth-core==1.2.0
social-auth-app-django==1.1.0
pytz==2017.02
raven==6.0.0
requests==2.13
six==1.10.0
hiredis==0.2.0